[MOD] Seperate UT and IT
